Refactor room components
