+++ app/client/components/rooms/RoomCards.tsx
@@ -1,31 +1,15 @@
-import { RoomAPIData, zosiaApi, zosiaApiRoutes } from "@client/utils/zosiaApi";
+import {
+  convertRoomAPIDataToRoomData,
+  RoomAPIData,
+  RoomData,
+} from "@client/utils/roomData";
+import { zosiaApi, zosiaApiRoutes } from "@client/utils/zosiaApi";
 import { Context } from "@reactivated";
 import { useQuery } from "@tanstack/react-query";
-import { parseISO } from "date-fns";
 import React, { useContext } from "react";
 import { Alert } from "../alert/Alert";
 import { ApiErrorMessage } from "./ApiErrorMessage";
 import { RoomCard } from "./RoomCard";
-
-export interface RoomData {
-  id: number;
-  name: string;
-  description: string;
-  members: RoomMember[];
-  lock?: {
-    user: RoomMember;
-    password?: string;
-    expirationDate: Date;
-  };
-  availableBedsSingle: number;
-  availableBedsDouble: number;
-}
-
-export interface RoomMember {
-  id: number;
-  firstName: string;
-  lastName: string;
-}
 
 interface RoomCardsProps {
   /** This is initial room data, used during server-side rendering. */
@@ -38,32 +22,8 @@ export const RoomCards = ({ initialRoomData }: RoomCardsProps) => {
   const { isPending, isError, data, error } = useQuery({
     queryKey: [zosiaApiRoutes.rooms],
     queryFn: async () => {
-      const res = await zosiaApi.get(zosiaApiRoutes.rooms);
-      const rooms = res.data as RoomAPIData[];
-
-      return rooms.map((room) => ({
-        id: room.id,
-        name: room.name,
-        description: room.description,
-        members: room.members.map((member) => ({
-          id: member.user.id,
-          firstName: member.user.first_name,
-          lastName: member.user.last_name,
-        })),
-        lock: room.lock
-          ? {
-              user: {
-                id: room.lock.user.id,
-                firstName: room.lock.user.first_name,
-                lastName: room.lock.user.last_name,
-              },
-              password: room.lock.password ?? undefined,
-              expirationDate: parseISO(room.lock.expiration_date),
-            }
-          : undefined,
-        availableBedsSingle: room.available_beds_single,
-        availableBedsDouble: room.available_beds_double,
-      }));
+      const res = await zosiaApi.get<RoomAPIData[]>(zosiaApiRoutes.rooms);
+      return res.data.map(convertRoomAPIDataToRoomData);
     },
     initialData: initialRoomData,
     // With SSR, set staleTime above 0 to avoid refetching immediately on the client
