+++ app/client/components/rooms/RoomMutations.tsx
@@ -1,0 +1,85 @@
+import { RoomAPIData } from "@client/utils/roomData";
+import { zosiaApi, zosiaApiRoutes } from "@client/utils/zosiaApi";
+import { Context } from "@reactivated";
+import { useMutation, useQueryClient } from "@tanstack/react-query";
+import React, { useContext } from "react";
+import { showCustomToast } from "../CustomToast";
+import { ApiErrorMessage } from "./ApiErrorMessage";
+
+export const useRoomMutations = (roomId: number, roomName: string) => {
+  const { user } = useContext(Context);
+
+  const queryClient = useQueryClient();
+
+  const onMutationSuccess = (message: string) => {
+    queryClient.invalidateQueries({
+      queryKey: [zosiaApiRoutes.rooms],
+    });
+    showCustomToast("success", message);
+  };
+
+  const onMutationError = (error: Error) => {
+    showCustomToast("error", <ApiErrorMessage error={error} />);
+    console.error(error);
+  };
+
+  const joinRoomMutation = useMutation({
+    mutationFn: async (password?: string) => {
+      return await zosiaApi.post<RoomAPIData>(
+        zosiaApiRoutes.roomMember(roomId),
+        {
+          user: user.id,
+          password: password,
+        },
+      );
+    },
+    onSuccess: () => onMutationSuccess(`You've joined room ${roomName}.`),
+    onError: onMutationError,
+  });
+
+  const leaveRoomMutation = useMutation({
+    mutationFn: async () => {
+      return await zosiaApi.delete<RoomAPIData>(
+        zosiaApiRoutes.roomMember(roomId),
+        {
+          data: { user: user.id },
+        },
+      );
+    },
+    onSuccess: () => onMutationSuccess(`You've left room ${roomName}.`),
+    onError: onMutationError,
+  });
+
+  const lockRoomMutation = useMutation({
+    mutationFn: async () => {
+      return await zosiaApi.post<RoomAPIData>(zosiaApiRoutes.lockRoom(roomId), {
+        user: user.id,
+      });
+    },
+    onSuccess: () =>
+      onMutationSuccess(
+        `You've locked room ${roomName}. Share the password with your friends.`,
+      ),
+    onError: onMutationError,
+  });
+
+  const unlockRoomMutation = useMutation({
+    mutationFn: async () => {
+      return await zosiaApi.delete<RoomAPIData>(
+        zosiaApiRoutes.lockRoom(roomId),
+      );
+    },
+    onSuccess: () =>
+      onMutationSuccess(
+        `You've unlocked room ${roomName}. Now everybody can join it.`,
+      ),
+    onError: onMutationError,
+  });
+
+  return {
+    joinRoomMutation,
+    leaveRoomMutation,
+    lockRoomMutation,
+    unlockRoomMutation,
+  };
+};
