+++ app/client/utils/zosiaApi.ts
@@ -1,29 +1,6 @@
 import { reverse } from "@reactivated";
 import axios from "axios";
 
-export interface RoomAPIData {
-  id: number;
-  name: string;
-  description: string;
-  members: {
-    user: RoomAPIUserData;
-    joined_at: string;
-  }[];
-  lock: {
-    user: RoomAPIUserData;
-    password: string | null;
-    expiration_date: string;
-  } | null;
-  available_beds_single: number;
-  available_beds_double: number;
-}
-
-interface RoomAPIUserData {
-  id: number;
-  first_name: string;
-  last_name: string;
-}
-
 export const zosiaApi = axios.create({
   // '/' in baseURL resolves to the current domain
   // so for example in development it will be http://localhost:8000/
